requirements.txt
amqp==2.4.2
backports.csv==1.0.7
billiard==3.6.0.0
celery==4.3.0
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2.1
django-cors-headers==2.5.3
django-crispy-forms==1.7.2
django-excel==0.0.10
django-heroku==0.3.1
django-import-export==1.2.0
djangorestframework==3.9.3
et-xmlfile==1.0.1
gunicorn==19.9.0
jdcal==1.4.1
kombu==4.5.0
lml==0.0.9
odfpy==1.4.0
openpyxl==2.6.2
psycopg2==2.8.2
pyexcel==0.5.13
pyexcel-io==0.5.17
pyexcel-webio==0.1.4
pytz==2019.1
PyYAML==5.1
sqlparse==0.3.0
tablib==0.13.0
texttable==1.6.1
vine==1.3.0
whitenoise==4.1.2
xlrd==1.2.0
xlwt==1.3.0
